Initial Commit
The first (and maybe only) commit of this repository - I coded everything offline so I don't have version control or any changes I can think to make at the moment. Right now, this creates a simple interface for Whisper by offering two options: having a "rolling" time limit and transcribing the files in batches, and having a "indefinite" time limit and listening for an F8 keypress to end the recording and begin transcribing data. Both work flawlessly, and Whisper takes roughly 2 seconds to transcribe anything from 5-60 seconds of data (I haven't tested it beyond that), so the rolling option can be set to cycle on anything from 4 seconds to as high as you want.

The only issue I can think of currently is that if you use the rolling method, once the duration is up and another loop begins, Whisper interprets that as the end of your sentence and doesn't merge the two (your previous and current transcriptions) very well. It may be possible to transcribe it live, however, but I haven't looked into it.